(struct _fise_stream.offset): Change type to off_t

+++ mailbox/file_stream.c
@@ -46,7 +46,7 @@
 struct _file_stream
 {
   FILE *file;
-  int offset;
+  off_t offset;
 
   char *filename;
   mu_stream_t cache;
